updated all the delete routes to delete the images

+++ src/routes/images/deleteImage.ts
@@ -6,7 +6,8 @@ import authenticateToken from "../../authRoute/authentication";
 // Get all images
 router.delete("/all", authenticateToken, async (req, res) => {
   try {
-    const images = await Image.deleteMany({});
+    const owner = req.body.user;
+    const images = await Image.deleteMany({ owner });
     res.status(201).send("Deleted all the images from the repository!");
   } catch (err) {
     res.status(500).json({ message: err.message });
@@ -16,8 +17,9 @@ router.delete("/all", authenticateToken, async (req, res) => {
 // get By ID:
 router.delete("/:id", authenticateToken, async (req, res) => {
   try {
-    const image = await Image.findById(req.params.id);
-    await Image.findByIdAndRemove(req.params.id);
+    const owner = req.body.user;
+    const _id = req.params.id;
+    const image = await Image.findOneAndDelete({ owner, _id });
     res.status(201).json(image);
   } catch (err) {
     return res.status(500).json({ message: err.message });
